refactor(wise_feedback): reorganize src/ui into screens, widgets and theme
