+++ packages/wise_feedback/lib/src/screens/screens.dart
@@ -1,0 +1,1 @@
+export 'feedback_form.dart';
