+++ packages/wise_feedback/lib/src/theme/theme.dart
@@ -1,0 +1,1 @@
+export 'wise_feedback_theme.dart';
