+++ packages/wise_feedback/lib/src/widgets/widgets.dart
@@ -1,5 +1,3 @@
 export 'feedback_button.dart';
-export 'feedback_form.dart';
 export 'feedback_toast.dart';
 export 'linear_feedback.dart';
-export 'wise_feedback_theme.dart';
